+++ src/Support/Changelog.md
@@ -6,6 +6,7 @@
 ##### 2024-XX-YY
 
 - Changed the `ConfigurableModel::hasConfiguration()` trait to return false on empty arrays as well (until now, only `null` was considered as no config)
+- Added the search engine and inventory feature configuration wrappers
 
 ## 4.1.0
 ##### 2024-07-11